delete model tests and move models to parent folder
